requirements.txt
anthropic>=0.40.0
openai>=1.50.0
python-dotenv>=1.0.0
httpx>=0.27.0
pydantic>=2.0.0
fastapi>=0.110.0
uvicorn>=0.23.0
PyPDF2>=3.0.0
chromadb>=0.5.0
PyPDF2
deepeval

